mmsegmentation/requirements.txt
-r requirements/optional.txt
-r requirements/runtime.txt
-r requirements/tests.txt
-r requirements/multimodal.txt

mmengine~=0.10.3
opencv-python~=4.8.1.78
mmcv~=2.1.0
pillow~=10.1.0
fvcore~=0.1.5.post20221221
dcnv4~=1.0.0.post2
mmdet~=3.3.0
requests~=2.28.2
seaborn~=0.13.2
torchvision~=0.17.1
tqdm~=4.65.2
scikit-learn~=1.3.2
pandas~=2.1.3
setuptools~=60.2.0
yaml~=0.2.5
pyyaml~=6.0.1
oss2~=2.17.0